feat: scraped previous design and remade james design utilising tailwind
I scraped the previous because of a worse user experience and non-fluent js
