+++ src/routes/_layout.svelte
@@ -3,8 +3,7 @@
 	import Tailwind from '../components/atoms/Tailwind.svelte';
 	import Styles from '../components/atoms/Styles.svelte';
 	import Footer from '../components/organisms/Footer.svelte';
-	import OldNav from '../components/oldNav.svelte';
-	import ReferenceNav from '../components/referenceNav.svelte';
+	import Nav from '../components/organisms/Nav.svelte';
 </script>
 
 <svelte:head>
@@ -13,7 +12,6 @@
 
 <Tailwind />
 <Styles />
-<ReferenceNav />
-<OldNav />
+<Nav />
 <slot />
 <Footer />
