+++ tailwind.config.js
@@ -35,6 +35,7 @@ module.exports = {
 				'7xl': '5rem',
 				'8xl': '6rem',
 				'9xl': '7rem',
+				'xss': '0.6rem',
 			},
 			height: {
 				'1/4': '25%',
